SDL_RWops implementation from int to long int in stream.cpp

+++ src/shared/stream.cpp
@@ -479,20 +479,20 @@ int listfiles(const char *dir, const char *ext, vector<char *> &files)
 }
 
 #ifndef STANDALONE
-static int rwopsseek(SDL_RWops *rw, int pos, int whence)
+static long int rwopsseek(SDL_RWops *rw, long int pos, int whence)
 {
     stream *f = (stream *)rw->hidden.unknown.data1;
     if((!pos && whence==SEEK_CUR) || f->seek(pos, whence)) return (int)f->tell();
     return -1;
 }
 
-static int rwopsread(SDL_RWops *rw, void *buf, int size, int nmemb)
+static long unsigned int rwopsread(SDL_RWops *rw, void *buf, long unsigned int size, long unsigned int nmemb)
 {
     stream *f = (stream *)rw->hidden.unknown.data1;
     return f->read(buf, size*nmemb)/size;
 }
 
-static int rwopswrite(SDL_RWops *rw, const void *buf, int size, int nmemb)
+static long unsigned int rwopswrite(SDL_RWops *rw, const void *buf, long unsigned int size, long unsigned int nmemb)
 {
     stream *f = (stream *)rw->hidden.unknown.data1;
     return f->write(buf, size*nmemb)/size;
